Refactor configuration.
