+++ GrandCentralBoard/Core/Configuration/WidgetSettings.swift
@@ -1,0 +1,29 @@
+//
+//  Created by Oktawian Chojnacki on 23.03.2016.
+//  Copyright © 2016 Oktawian Chojnacki. All rights reserved.
+//
+
+import Decodable
+
+
+public struct WidgetSettings {
+    public let name: String
+    public let settings: AnyObject
+
+    public init(name: String, settings: AnyObject) {
+        self.name = name
+        self.settings = settings
+    }
+
+    @warn_unused_result public static func settingsFromArray(array: [AnyObject]) -> [WidgetSettings] {
+        return array.flatMap({ settings in
+            return try? WidgetSettings.decode(settings)
+        })
+    }
+}
+
+extension WidgetSettings : Decodable {
+    public static func decode(jsonObject: AnyObject) throws -> WidgetSettings {
+        return try WidgetSettings(name: jsonObject => "name", settings: jsonObject => "settings")
+    }
+}
